encoder is now working correctly, i.e. fixed bug when going from an to 1000 and 1 value. But its not a pretty solution..

+++ Code/lib/radio_encoder/radio_encoder.cpp
@@ -62,7 +62,7 @@ void rotary_onButtonClick() {
     button_clicked = true;*/
 }
 
-int rotary_loop() {
+int16_t rotary_loop() {
 
 
 	//dont print anything unless value changed
